requirement.txt
streamlit
python-dotenv
PyPDF2
python-docx
google-genai
pytest
black
isort
flake8
